requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiohttp-cors==0.7.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorful==0.5.6
distlib==0.3.8
fastapi==0.112.0
filelock==3.15.4
frozenlist==1.4.1
google-api-core==2.19.1
google-auth==2.33.0
googleapis-common-protos==1.63.2
grpcio==1.65.4
h11==0.14.0
httptools==0.6.1
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-py-plugins==0.4.1
mdurl==0.1.2
memray==1.13.4
msgpack==1.0.8
multidict==6.0.5
opencensus==0.11.4
opencensus-context==0.1.3
packaging==24.1
platformdirs==4.2.2
prometheus_client==0.20.0
proto-plus==1.24.0
protobuf==5.27.3
py-spy==0.3.14
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
ray==2.34.0
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
starlette==0.37.2
textual==0.75.1
typing_extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0
virtualenv==20.26.3
watchfiles==0.23.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
numpy==1.26.4
